modularize docker watcher
