+++ src/tracer/src/extracts/containers/docker_watcher/mod.rs
@@ -1,0 +1,2 @@
+pub mod event;
+pub(super) mod watcher;
